requirements.txt
requests
playwright
qrcode
pillow
Appium-Python-Client
selenium
